requirements.txt
openai~=0.27.8
gradio~=3.35.2
python-dotenv~=1.0.0